Revert "Rv activity"
