+++ app/src/main/AndroidManifest.xml
@@ -13,12 +13,10 @@
         android:theme="@style/Theme.WhatsAppClone"
         tools:targetApi="31">
         <activity
-            android:name=".rvActivity.RvMainActivity"
-            android:exported="false" />
-        <activity
-            android:name=".llActivity.MainActivity"
+            android:name=".MainActivity"
+            android:theme="@style/SplashTheme"
             android:exported="true"
-            android:theme="@style/SplashTheme">
+            >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
