+++ app/src/main/java/com/example/whatsappclone/Chat.kt
@@ -1,4 +1,6 @@
-package com.example.whatsappclone.data
+package com.example.whatsappclone
+
+import java.util.Date
 
 data class Chat(
     val profilePhoto: Int,
@@ -6,6 +8,7 @@ data class Chat(
     val message: String,
     val messageDate: String,
     val messageCount: String
+
 )
