--- app/src/main/java/com/example/whatsappclone/viewModelChat/ChatIUState.kt
@@ -1,7 +1,0 @@
-package com.example.whatsappclone.viewModelChat
-
-import com.example.whatsappclone.data.Chat
-
-data class ChatIUState(
-    val chats: List<Chat> = emptyList()
-)
