+++ app/src/main/res/layout/chats_screen.xml
@@ -3,65 +3,66 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/chatsScreen"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="horizontal"
-    android:padding="10dp">
+    android:layout_height="match_parent">
+
+    <!-- android:background="@drawable/profile_shape"-->
 
     <de.hdodenhof.circleimageview.CircleImageView
         android:id="@+id/imageProfilePicture"
         android:layout_width="50dp"
         android:layout_height="50dp"
-        android:layout_gravity="center_vertical"
-        android:layout_marginEnd="10dp"
+        android:layout_centerInParent="true"
+        android:layout_marginStart="10dp"
+        android:layout_marginTop="20dp"
         android:scaleType="centerCrop"
         android:src="@drawable/media_photo" />
 
     <LinearLayout
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_vertical"
-        android:layout_weight="1"
-        android:orientation="vertical"
-        android:paddingEnd="10dp">
+        android:layout_width="230dp"
+        android:layout_height="40dp"
+        android:layout_marginStart="10dp"
+        android:layout_marginTop="25dp"
+        android:layout_marginEnd="10dp"
+        android:orientation="vertical">
 
         <TextView
             android:id="@+id/chatName"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
+            android:layout_width="200dp"
+            android:layout_height="20dp"
             android:textStyle="bold"
             android:textSize="15sp"
-            android:text="@string/chat_name" />
+            android:text="Me" />
 
         <TextView
             android:id="@+id/message"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/message" />
+            android:layout_width="200dp"
+            android:layout_height="20dp"
+            android:text="Hey sup" />
     </LinearLayout>
 
     <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_vertical"
-        android:orientation="vertical"
-        android:gravity="end">
+        android:layout_width="90dp"
+        android:layout_height="40dp"
+        android:layout_marginTop="25dp"
+        android:layout_marginEnd="10dp"
+        android:orientation="vertical">
 
         <TextView
             android:id="@+id/date"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
+            android:layout_width="85dp"
+            android:layout_height="20dp"
             android:textAlignment="textEnd"
             android:textColor="@color/GreenWhatsApp"
-            android:text="@string/date" />
+            android:text="8/6/2024" />
 
         <TextView
             android:id="@+id/numberChats"
             android:layout_width="20dp"
             android:layout_height="20dp"
+            android:layout_marginStart="65dp"
             android:textAlignment="center"
-            android:layout_marginTop="5dp"
-            android:background="@drawable/layout_message_count"
             android:textColor="@color/white"
-            android:text="@string/numberChats" />
+            android:background="@drawable/layout_message_count"
+            android:text="1" />
     </LinearLayout>
 </LinearLayout>
