+++ app/src/main/res/layout/footer_chat_menu.xml
@@ -1,0 +1,124 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/footer_chat_menu"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+        <ImageButton
+            android:id="@+id/chatsButton"
+            android:layout_width="100dp"
+            android:layout_height="24dp"
+            android:layout_marginStart="5dp"
+            android:background="@null"
+            android:scaleType="centerInside"
+            android:foregroundGravity="center"
+            android:layout_weight="0.1"
+            app:srcCompat="@drawable/chats_button_icon"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintBottom_toTopOf="@id/chatText"
+            />
+
+        <TextView
+            android:id="@+id/chatText"
+            android:layout_width="100dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="5dp"
+            android:layout_marginBottom="20dp"
+            android:layout_weight="0.1"
+            android:gravity="center"
+            android:hint="Chats"
+            android:textSize="16sp"
+            android:fontFamily="sans-serif"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            />
+
+        <ImageButton
+            android:id="@+id/updatesButton"
+            android:layout_width="100dp"
+            android:layout_height="24dp"
+            android:layout_marginStart="5dp"
+            android:background="@null"
+            android:scaleType="centerInside"
+            android:foregroundGravity="center"
+            android:layout_weight="0.1"
+            app:srcCompat="@drawable/updates_button_icon"
+            app:layout_constraintBottom_toTopOf="@id/updateText"
+            app:layout_constraintStart_toEndOf="@id/chatsButton"
+            />
+
+        <TextView
+            android:id="@+id/updateText"
+            android:layout_width="100dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="5dp"
+            android:layout_marginBottom="20dp"
+            android:layout_weight="0.1"
+            android:gravity="center"
+            android:hint="Updates"
+            android:textSize="16sp"
+            android:fontFamily="sans-serif"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toEndOf="@id/chatText"
+            />
+
+        <ImageButton
+            android:id="@+id/communitiesButton"
+            android:layout_width="100dp"
+            android:layout_height="24dp"
+            android:layout_marginStart="5dp"
+            android:background="@null"
+            android:scaleType="centerInside"
+            android:foregroundGravity="center"
+            android:layout_weight="0.1"
+            app:srcCompat="@drawable/communities_icon"
+            app:layout_constraintBottom_toTopOf="@id/communitiesText"
+            app:layout_constraintStart_toEndOf="@id/updatesButton"
+            />
+
+        <TextView
+            android:id="@+id/communitiesText"
+            android:layout_width="100dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="5dp"
+            android:layout_marginBottom="20dp"
+            android:layout_weight="0.1"
+            android:gravity="center"
+            android:hint="Communities"
+            android:textSize="16sp"
+            android:fontFamily="sans-serif"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toEndOf="@id/updateText"
+            />
+
+        <ImageButton
+            android:id="@+id/callButton"
+            android:layout_width="50dp"
+            android:layout_height="24dp"
+            android:layout_marginStart="20dp"
+            android:background="@null"
+            android:scaleType="centerInside"
+            android:foregroundGravity="center"
+            android:layout_weight="0.1"
+            app:srcCompat="@drawable/calls_icon"
+            app:layout_constraintBottom_toTopOf="@id/callsText"
+            app:layout_constraintStart_toEndOf="@id/communitiesButton"
+            />
+
+        <TextView
+            android:id="@+id/callsText"
+            android:layout_width="50dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="20dp"
+            android:layout_marginBottom="20dp"
+            android:layout_weight="0.1"
+            android:gravity="center"
+            android:hint="Calls"
+            android:textSize="16sp"
+            android:fontFamily="sans-serif"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toEndOf="@id/communitiesText"
+            />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
